+++ WPFHospitalEditor/AllConstants.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO;
 
-namespace WPFHospitalEditor.MapObjectModel
+namespace WPFHospitalEditor
 {
 
     public static class AllConstants
@@ -9,5 +9,7 @@ namespace WPFHospitalEditor.MapObjectModel
         public const double doorHeight = 6;
         public static char separator = Path.DirectorySeparatorChar;
         public readonly static string MAPOBJECT_PATH = $"..{separator}..{separator}..{separator}..{separator}WPFHospitalEditor{separator}Repository{separator}Data{separator}AllMapObjects.json";
+        public const int additionalInformationsbuttonWidth = 100;
+        public const int additionalInformationsbuttonHeight = 25;
     }
 }
